requirements.txt
tensorflow
gradio
numpy
Pillow
scikit-learn
matplotlib
seaborn
